Add EnumIntoGetters and EnumAsGetters
`into_*` methods and `as_*` methods can be generated seperately from
each other using the aforemented derives. `EnumGetters` is no longer the
name of a derive, so don't use that.

Copy types are no longer being tracked, so we don't need the lazy_static
crate anymore (which was the only thing using it).

Tests for `as_*` getters have been updated.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>